Factor out new test utils
