+++ src/editor/nodes/collapsibleBlock/keyboardShortcutCommands/backspace.test.ts
@@ -1,38 +1,9 @@
 import { Editor } from '@tiptap/react';
 
 import { defaultCollapsibleBlock, uncollapsedCollapsibleBlockWithEmptyContent } from '../../../test-fixtures';
-import { findNodesByNodeType, getText } from '../../../test-utils';
+import { findNodesByNodeType, getPrettyHTML, getText, setUpEditorWithSelection } from '../../../test-utils';
 import { EDITOR_EXTENSIONS } from '../../../TipTapEditorConfigs';
 import { backspace } from './backspace';
-
-const collapsibleBlockWithSelection = `
-<collapsible-block>
-    <collapsible-summary>He|ader|</collapsible-summary>
-    <collapsible-content>
-        <p>Content Line 1</p>
-        <p>Content Line 2</p>
-    </collapsible-content>
-</collapsible-block>`;
-
-function setUpEditorWithSelection(editor: Editor, content: string) {
-  editor.chain().setContent(content).run();
-  const docLength = editor.getText().length;
-  const positions = [];
-  for (let i = 0; i < docLength; i++) {
-    const text = editor.view.state.doc.textBetween(i, i + 1);
-    if (text === '|') {
-      positions.push(i - positions.length);
-    }
-  }
-  expect(positions.length).toBeGreaterThanOrEqual(1);
-  expect(positions.length).toBeLessThanOrEqual(2);
-  editor
-    .chain()
-    .setContent(content.replaceAll('|', ''))
-    .setTextSelection(positions.length > 1 ? { from: positions[0], to: positions[1] } : positions[0])
-    .run();
-  return positions;
-}
 
 describe('Backspace keyboard shortcut command', () => {
   const editor = new Editor({
@@ -40,10 +11,16 @@ describe('Backspace keyboard shortcut command', () => {
   });
 
   it('should select text based on markers', () => {
-    const [from, to] = setUpEditorWithSelection(editor, collapsibleBlockWithSelection);
-    console.log(from, to);
-    editor.chain().setTextSelection({ from, to }).run();
-
+    const [from, to] = setUpEditorWithSelection(
+      editor,
+      `<collapsible-block>
+        <collapsible-summary>He|ader|</collapsible-summary>
+        <collapsible-content>
+            <p>Content Line 1</p>
+            <p>Content Line 2</p>
+        </collapsible-content>
+      </collapsible-block>`,
+    );
     const sel = editor.view.state.selection;
     const text = editor.view.state.doc.textBetween(sel.from, sel.to);
     expect(sel.from).toEqual(from);
@@ -52,10 +29,7 @@ describe('Backspace keyboard shortcut command', () => {
   });
 
   test('should turn collapsible block back into paragraphs when selection is at the beginning of the summary', () => {
-    // 3 is to position the caret at the beginning of the summary
-    // editor.chain().setContent(collapsibleBlockWithSelection).run();
-    // editor.chain().setContent(collapsibleBlockWithSelection).setTextSelection({ from: 3, to: 5 }).run();
-    const [pos] = setUpEditorWithSelection(
+    setUpEditorWithSelection(
       editor,
       `<collapsible-block>
         <collapsible-summary>|Header</collapsible-summary>
@@ -65,7 +39,6 @@ describe('Backspace keyboard shortcut command', () => {
         </collapsible-content>
       </collapsible-block>`,
     );
-    expect(pos).toEqual(3);
 
     expect(editor.state.doc.childCount).toBe(1);
 
@@ -74,14 +47,20 @@ describe('Backspace keyboard shortcut command', () => {
     const commandResult = backspace({ editor });
     expect(commandResult).toBe(true);
 
-    const { doc } = editor.state;
-    console.log(editor.getHTML());
-    // A new collapsible should have been added
-    expect(doc.childCount).toBe(3);
-
-    expect(getText(doc.child(0))).toEqual('Header');
-    expect(getText(doc.child(1))).toEqual('Content Line 1');
-    expect(getText(doc.child(2))).toEqual('Content Line 2');
+    expect(getPrettyHTML(editor)).toMatchInlineSnapshot(`
+      "<>
+        <draggable-block>
+          <p>Header</p>
+        </draggable-block>
+        <draggable-block>
+          <p>Content Line 1</p>
+        </draggable-block>
+        <draggable-block>
+          <p>Content Line 2</p>
+        </draggable-block>
+      </>;
+      "
+    `);
   });
 
   test('should remove content and collapse block when removing the only remaining content block of collapsible block', () => {
